fix(ci): resolve all failing GitHub Actions checks
CI / lint-and-format:
- Migrate deprecated top-level ruff lint config to [tool.ruff.lint]
- Fix all ruff check errors (imports, exception chaining, unused vars)
- Apply ruff format across backend (55 files)
- Achieve clean mypy run: explicit_package_bases fixes alembic dual-module
  error; fix genuine bugs (missing json/text/ClassVar/Any imports,
  StreamProducer.publish trace_id TypeError, neo4j None guards);
  targeted overrides for tests and third-party stub gaps

CI / unit-tests & integration-tests:
- Add cryptography + python-multipart runtime deps, respx dev dep
- Fix MinioContainer class name (was MinIOContainer)
- Guard bentoml service import with graceful skip in extraction graph test
- Add missing test imports (mock, respx, asyncio, numpy, Neo4jClient)

CI / frontend-lint:
- Generate package-lock.json; add tsconfig.json and .eslintrc.json

Security Audit / pip-audit:
- Create backend/requirements.txt (mirrors pyproject deps)

Security Audit / trufflehog:
- Remove base/head inputs that made BASE==HEAD on push events

+++ backend/db/vector_store.py
@@ -10,9 +10,11 @@ security.
 
 from __future__ import annotations
 
+import json
 from typing import Any
 
 import structlog
+from sqlalchemy import text
 
 logger = structlog.get_logger(__name__)
 
@@ -26,7 +28,7 @@ DEFAULT_ALPHA = 0.5
 DEFAULT_TOP_K = 10
 
 
-def rrf_score(ranks: list[int], k: int = RRF_K) -> float:
+def rrf_score(ranks: list[float], k: int = RRF_K) -> float:
     """Compute the Reciprocal Rank Fusion score for a document.
 
     Given the *ranks* a document received across multiple rankings (e.g. one
